+++ src/applications/vaos/new-appointment/components/ClinicChoicePage/index.jsx
@@ -1,21 +1,24 @@
 import React, { useEffect } from 'react';
-import { shallowEqual, useDispatch, useSelector } from 'react-redux';
+import { useHistory } from 'react-router-dom';
+import { useDispatch, useSelector } from 'react-redux';
 import SchemaForm from 'platform/forms-system/src/js/components/SchemaForm';
 import FormButtons from '../../../components/FormButtons';
 import RequestEligibilityMessage from './RequestEligibilityMessage';
 import FacilityAddress from '../../../components/FacilityAddress';
 import { scrollAndFocus } from '../../../utils/scrollAndFocus';
-import { MENTAL_HEALTH, PRIMARY_CARE } from '../../../utils/constants';
 import {
   routeToNextAppointmentPage,
   routeToPreviousAppointmentPage,
 } from '../../redux/actions';
-import useFormState from '../../../hooks/useFormState';
-import { getSiteIdFromFacilityId } from '../../../services/location';
-import { getClinicId } from '../../../services/healthcare-service';
 
-import { getClinicPageInfo } from '../../redux/selectors';
-import { useHistory } from 'react-router-dom';
+import {
+  selectTypeOfCare,
+  selectPageChangeInProgress,
+  selectChosenFacilityInfo,
+  selectEligibility,
+} from '../../redux/selectors';
+import useClinicFormState from './useClinicFormState';
+import { MENTAL_HEALTH, PRIMARY_CARE } from '../../../utils/constants';
 
 function formatTypeOfCare(careLabel) {
   if (careLabel.startsWith('MOVE') || careLabel.startsWith('CPAP')) {
@@ -39,132 +42,39 @@ function getPageTitle(schema, typeOfCare, usingPastClinics) {
   }
   return pageTitle;
 }
-const initialSchema = {
-  type: 'object',
-  required: ['clinicId'],
-  properties: {
-    clinicId: {
-      type: 'string',
-      enum: [],
-    },
-  },
-};
-const uiSchema = {
-  clinicId: {
-    'ui:widget': 'radio',
-  },
-};
+
 const pageKey = 'clinicChoice';
 export default function ClinicChoicePage() {
-  const {
-    data: initialData,
-    canMakeRequests,
-    clinics,
-    eligibility,
-    facility,
-    pageChangeInProgress,
-    pastAppointments,
-    typeOfCare,
-  } = useSelector(state => getClinicPageInfo(state, pageKey), shallowEqual);
   const dispatch = useDispatch();
   const history = useHistory();
-  const typeOfCareLabel = formatTypeOfCare(typeOfCare.name);
-  const { data, schema, setData } = useFormState({
-    initialSchema() {
-      let newSchema = initialSchema;
-      let filteredClinics = clinics;
 
-      if (pastAppointments) {
-        const pastAppointmentDateMap = new Map();
-        const siteId = getSiteIdFromFacilityId(initialData.vaFacility);
+  const facility = useSelector(selectChosenFacilityInfo);
+  const typeOfCare = useSelector(selectTypeOfCare);
+  const pageChangeInProgress = useSelector(selectPageChangeInProgress);
+  const eligibility = useSelector(selectEligibility);
 
-        pastAppointments.forEach(appt => {
-          const apptTime = appt.startDate;
-          const latestApptTime = pastAppointmentDateMap.get(appt.clinicId);
-          if (
-            // Remove parse function when converting the past appointment call to FHIR service
-            appt.facilityId === siteId &&
-            (!latestApptTime || latestApptTime > apptTime)
-          ) {
-            pastAppointmentDateMap.set(appt.clinicId, apptTime);
-          }
-        });
-
-        filteredClinics = clinics.filter(clinic =>
-          // Get clinic portion of id
-          pastAppointmentDateMap.has(getClinicId(clinic)),
-        );
-      }
-
-      if (filteredClinics.length === 1) {
-        const clinic = filteredClinics[0];
-        newSchema = {
-          ...newSchema,
-          properties: {
-            clinicId: {
-              type: 'string',
-              title: `Would you like to make an appointment at ${
-                clinic.serviceName
-              }?`,
-              enum: [clinic.id, 'NONE'],
-              enumNames: [
-                'Yes, make my appointment here',
-                'No, I need a different clinic',
-              ],
-            },
-          },
-        };
-      } else {
-        newSchema = {
-          ...newSchema,
-          properties: {
-            clinicId: {
-              type: 'string',
-              title:
-                'Choose a clinic below or request a different clinic for this appointment.',
-              enum: filteredClinics.map(clinic => clinic.id).concat('NONE'),
-              enumNames: filteredClinics
-                .map(clinic => clinic.serviceName)
-                .concat('I need a different clinic'),
-            },
-          },
-        };
-      }
-
-      return newSchema;
-    },
+  const {
+    data,
+    schema,
     uiSchema,
-    initialData,
-  });
+    setData,
+    firstMatchingClinic,
+  } = useClinicFormState();
 
+  const typeOfCareLabel = formatTypeOfCare(typeOfCare.name);
   const usingUnsupportedRequestFlow =
-    data.clinicId === 'NONE' && !canMakeRequests;
+    data.clinicId === 'NONE' && !eligibility?.request;
   const usingPastClinics =
     typeOfCare.id !== PRIMARY_CARE && typeOfCare.id !== MENTAL_HEALTH;
 
-  // useEffect(() => {
-  //   dispatch(openClinicPage(pageKey, uiSchema, initialSchema));
-  // }, []);
-
-  useEffect(
-    () => {
-      scrollAndFocus();
-      document.title = `${getPageTitle(
-        schema,
-        typeOfCare,
-        usingPastClinics,
-      )} | Veterans Affairs`;
-    },
-    [usingPastClinics],
-  );
-
-  // if (!schemaAndFacilityReady) {
-  //   return <LoadingIndicator message="Loading your facility and clinic info" />;
-  // }
-
-  const firstMatchingClinic = clinics?.find(
-    clinic => clinic.id === schema?.properties.clinicId.enum[0],
-  );
+  useEffect(() => {
+    scrollAndFocus();
+    document.title = `${getPageTitle(
+      schema,
+      typeOfCare,
+      usingPastClinics,
+    )} | Veterans Affairs`;
+  }, []);
 
   return (
     <div>
